Initializes a new VISION object.

Initializes a new VISION object.

Vision(data, ...)

# S4 method for matrixORSparse
Vision(data, signatures = list(),
  housekeeping = NULL, unnormalizedData = NULL, meta = NULL,
  nomodel = TRUE, projection_genes = c("fano"),
  min_signature_genes = 5, weights = NULL, threshold = 0.05,
  perm_wPCA = FALSE, projection_methods = c("tSNE30"),
  sig_norm_method = c("znorm_columns", "none", "znorm_rows",
  "znorm_rows_then_columns", "rank_norm_columns"),
  sig_score_method = c("naive", "weighted_avg"), pool = "auto",
  cellsPerPartition = 100, name = NULL, cluster_variable = "",
  latentSpace = NULL, latentTrajectory = NULL, pools = list())

# S4 method for data.frame
Vision(data, ...)

# S4 method for ExpressionSet
Vision(data, ...)

# S4 method for SummarizedExperiment
Vision(data, ...)

Arguments

data
expression data - can be one of these:

numeric matrix

ExpressionSet object

SummzrizedExperiment object (or extending classes)

signatures
list of file paths to signature files (.gmt or .txt) or a list of Signature objects

housekeeping
vector of gene names

meta
data file with cell meta data (.txt), or a data.frame with meta information. Note that rows should match samples in the data, and columns should either be factors or numerics.

nomodel
if TRUE, no fnr curve calculated and all weights equal to 1. Else FNR and weights calculated. [Default:TRUE]

projection_genes
name of method ('threshold' or 'fano') or list of genes to use when computing projections.

min_signature_genes
Minimum number of genes required to compute a signature

weights
Precomputed weights for each coordinate. Normally computed from the FNR curve.

threshold
Threshold to apply for the threshold filter. Number of cells or proportion of cells (if less than 1)

perm_wPCA
If TRUE, apply permutation WPCA to calculate significant number of PCs. Else not. Default FALSE.

projection_methods
a character list of which projection methods to apply. Can be:

tSNE10 (tSNE with perplexity 10)

tSNE30 (tSNE with perplexity 30)

ICA

ISOMap

RBFPCA

By default will perform tSNE and PCA on the data.

sig_norm_method
Method to apply to normalize the expression matrix before calculating signature scores. Valid options are: "znorm_columns" (default), "none", "znorm_rows", "znorm_rows_then_columns", or "rank_norm_columns"

sig_score_method
Method to apply when calculating signature scores. Either "naive" (default) or "weighted_avg"

pool
indicates whether or not to create supercells. Acceptable values are TRUE, FALSE, or 'auto', the last of which is the default and enables pooling if there are more than 15000 cells.

cellsPerPartition
the minimum number of cells to put into a cluster

name
a name for the sample - shown on the output report

cluster_variable
variable to use to denote clusters

latentSpace
latent space for expression data. Numeric matrix or dataframe with dimensions CELLS x COMPONENTS

latentTrajectory
trajectory to model cell progression. Wrapped result of a trajectory inference by the dynverse/dynwrap library

metaData
a dataframe storing all metadata (can be either factor or numerical), where the rows are cells and columns are individual meta data items

Value

A VISION object

Examples

# NOT RUN {
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

# choose housekeeping genes
hkg <- housekeeping$default

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- Vision(data = expMat,
                     signatures = sigs,
                     housekeeping = hkg)
# }