Pool cells into microclusters

Merges similar transcriptional profiles into representative 'pools'

applyMicroClustering(exprData, cellsPerPartition = 100,
  filterInput = "fano", filterThreshold = round(ncol(exprData) * 0.2),
  preserve_clusters = NULL, latentSpace = NULL)

Arguments

exprData
the expression data matrix

cellsPerPartition
control over the minimum number of cells to put into each supercell

filterInput
name of filtering method ('threshold' or 'fano') or list of genes to use when computing projections.

filterThreshold
Threshold to apply when using the 'threshold' projection genes filter. If greater than 1, this specifies the number of cells in which a gene must be detected for it to be used when computing PCA. If less than 1, this instead specifies the proportion of cells needed

preserve_clusters
named factor vector denoted cluster boundaries to preserve

latentSpace
(Optional) Latent space to be used instead of PCA numeric matrix cells x components

Value

pooled cells - named list of vectors - cells in each supercell

Details

A latent space is computed for the expression data via PCA after filtering on genes (using parameters filterInput and filterThreshold).

Alternately, a latent space can be supplied via the latentSpace argument

Euclidean distance within the latent space is then used to create cell pools