C++ wilcox rank-sums test

Given indices in cluster_num, and cluster_denom, compute the ranksums test on every row for values in cluster_num vs cluster_denom.

matrix_wilcox_cpp(
  data,
  cluster_num,
  cluster_denom,
  jobs = getOption("mc.cores", 1L)
)

Arguments

data

matrix of values, each row representing a separate variable

cluster_num

numeric vector - indices denoting the group to be compared

cluster_denom

numeric vector - indices denoting the other group to be compared

jobs

integer specifying number of parallel jobs. Can be used to override the default mc.cores option for when running this within an already-parallel loop

Value

pval - numeric vector, pvalue for each row stat - numeric vector, test statistic (AUC) for each row

Details

Ties are handled correctly but values are assumed to contain no NAs