Save the VISION object as an .RDS file and view the results on a localhost

Save the results object as an RDS file for future use, and launch a local server to explore the results with a browser.

# S4 method for Vision
saveAndViewResults(fpout, ofile = NULL, port = NULL,
  host = NULL, browser = TRUE, name = NULL)

Arguments

fpout
VISION object

ofile
the path to save the object in. If NULL, the object is saved in the working directory [default:NULL]

port
The port on which to serve the output viewer. If omitted, a random port between 8000 and 9999 is chosen.

host
The host used to serve the output viewer. If omitted, "127.0.0.1" is used.

Value

the path of the saved file

Examples

# NOT RUN {
expMat <- matrix(rnorm(200000), nrow=500)
rownames(expMat) <- paste0("gene",1:500)

#create 20 signatures of 25 genes each
sigs <- lapply(1:20, function(i) {
sigData <- sign(rnorm(25))
names(sigData) <- paste0("gene",sample(1:100,25))
return(createGeneSignature(name = paste0("sig",i),
                                 sigData = sigData))
})

fp <- Vision(data = expMat, signatures = sigs)

## analyze requires actual non-random data to run properly
# }# NOT RUN {
fp.out <- analyze(fp)
saveAndViewResults(fp.out)
# }