View results of analysis

launch a local server to explore the results with a browser.

# S4 method for Vision
viewResults(object, port = NULL, host = NULL,
  browser = TRUE, name = NULL)

# S4 method for character
viewResults(object, port = NULL, host = NULL,
  browser = TRUE, name = NULL)

Arguments

object
VISION object or path to an RDS file containing such an object (saved using saveAndViewResults, or directly using saveRDS)

port
The port on which to serve the output viewer. If omitted, a random port between 8000 and 9999 is chosen.

host
The host used to serve the output viewer. If omitted, "127.0.0.1" is used.

browser
Whether or not to launch the browser automatically (default=TRUE)

name
Name for the sample - is shown at the top of the output report

Value

None

Examples

# NOT RUN {
vis <- Vision(data = expMat, signatures = sigs)

options(mc.cores=10) # Use 10 cores
vis <- analyze(vis)

saveRDS(vis, 'vision_results.rds') # (Optional) Save results

viewResults(vis) # Launches dynamic output report
# }