Save the VISION object as an .RDS file and view the results on a localhost

This is just a convience function wrapping two function calls

# S4 method for Vision
saveAndViewResults(object, ofile = NULL,
  port = NULL, host = NULL, browser = TRUE, name = NULL)

Arguments

object
VISION object

ofile
the path to save the object in. If NULL, the object is saved in the working directory [default:NULL]

port
The port on which to serve the output viewer. If omitted, a random port between 8000 and 9999 is chosen.

host
The host used to serve the output viewer. If omitted, "127.0.0.1" is used.

browser
Whether or not to launch the browser automatically (default=TRUE)

name
Name for the sample - is shown at the top of the output report

Value

the path of the saved file

Details

saveAndViewResults(vis, 'vision_results.rds')

is equivalent to:

saveRDS(vis, 'vision_results.rds') viewResults(vis)

Examples

# NOT RUN {
vis <- Vision(data = expMat, signatures = sigs)

options(mc.cores=10) # Use 10 cores
vis <- analyze(vis)

saveAndViewResults(vis, 'vision_results.rds') # Saves and launches dynamic output report
# }